refactor: Break out data class
