+++ packages/app_center/lib/src/snapd/snap_model.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:app_center/snapd.dart';
+import 'package:app_center/src/snapd/snap_data.dart';
 import 'package:app_center/src/snapd/snapd_cache.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/widgets.dart';
@@ -55,6 +56,11 @@ final progressProvider =
   return streamController.stream;
 });
 
+@visibleForTesting
+final activeChangeIdProvider = StateProvider.family<String?, String>(
+  (ref, name) => null,
+);
+
 final activeChangeProvider =
     StreamProvider.family<SnapdChange, String?>((ref, id) {
   if (id == null) return const Stream.empty();
@@ -68,10 +74,6 @@ final activeChangeProvider =
   ref.onDispose(controller.close);
   return controller.stream;
 });
-
-final activeChangeIdProvider = StateProvider.family<String?, String>(
-  (ref, name) => null,
-);
 
 final snapInstallProvider =
     FutureProvider.family<void, String>((ref, snapName) async {
@@ -129,63 +131,6 @@ final snapRemoveProvider =
   final changeId = await getService<SnapdService>().remove(snapName);
   ref.read(activeChangeIdProvider(snapName).notifier).state = changeId;
 });
-
-@immutable
-class SnapData {
-  SnapData({
-    required this.name,
-    required this.localSnap,
-    required this.storeSnap,
-    this.activeChangeId,
-    String? selectedChannel,
-  }) : selectedChannel =
-            selectedChannel ?? _defaultSelectedChannel(localSnap, storeSnap);
-
-  final String name;
-  final Snap? localSnap;
-  final Snap? storeSnap;
-  final String? selectedChannel;
-  final String? activeChangeId;
-
-  Snap get snap => storeSnap ?? localSnap!;
-  SnapChannel? get channelInfo => storeSnap?.channels[selectedChannel];
-  bool get isInstalled => localSnap != null;
-  bool get hasGallery =>
-      storeSnap != null && storeSnap!.screenshotUrls.isNotEmpty;
-  Map<String, SnapChannel>? get availableChannels => storeSnap?.channels;
-
-  VoidCallback? callback(
-    WidgetRef ref,
-    SnapAction action, [
-    SnapLauncher? launcher,
-  ]) {
-    return switch (action) {
-      SnapAction.cancel => () => ref.read(snapAbortProvider(name)),
-      SnapAction.install =>
-        storeSnap != null ? () => ref.read(snapInstallProvider(name)) : null,
-      SnapAction.open =>
-        launcher?.isLaunchable ?? false ? launcher!.open : null,
-      SnapAction.remove => () => ref.read(snapRemoveProvider(name)),
-      SnapAction.switchChannel =>
-        storeSnap != null ? () => ref.read(snapRefreshProvider(this)) : null,
-      SnapAction.update =>
-        storeSnap != null ? () => ref.read(snapRefreshProvider(this)) : null,
-    };
-  }
-
-  static String? _defaultSelectedChannel(Snap? localSnap, Snap? storeSnap) {
-    final channels = storeSnap?.channels.keys;
-    final localChannel = localSnap?.trackingChannel;
-    if (localChannel != null && (channels?.contains(localChannel) ?? false)) {
-      return localChannel;
-    } else if (channels?.contains('latest/stable') ?? false) {
-      return 'latest/stable';
-    } else {
-      return channels?.firstWhereOrNull((c) => c.contains('stable')) ??
-          channels?.firstOrNull;
-    }
-  }
-}
 
 extension SnapdChangeX on SnapdChange {
   double get progress {
